AC4 - Refactor in ShortededLinesFormatter to get ready to reuse the highlight functionality and to organise it a bit better
